Renamed NodeConfig.NodeIP to NodeConfig.PodTrafficNodeIP
Prep work for segregating pod and management traffic.
To support traffic segregation, NodeConfig need to have multiple node IPs:
one for pod to pod traffic, master to node traffic, etc.
NodeConfig.NodeIP doesn't tell which traffic we are referring to.
NodeConfig.PodTrafficNodeIP is the node IP to be used for
pod to pod traffic.